+++ popup/templates/navbar.js
@@ -26,7 +26,6 @@ document.write(/*html*/`
     <!-- PROFILE -->
     <a href="profile.html" id="profile">
         <div class="progress-container" style="--level-progress: ${localStorage.getItem('level-progress') || 0}%">
-            <div class="progress-level-up-marker"></div>
             <img src="/images/wanikani-default.png">
         </div>
         <p title="Level">${localStorage.getItem("level") || ""}</p>